Drain connections for python3-http
When used with OpenFaaS Standard/Enterprise, the python3-http
template's handler will now ignore SIGTERM allowing the
watchdog and Kubernetes to handle the shutdown.

When there are ongoing requests, these will be processed
before exiting.

When there are no ongoing requests, the function will
exit immediately.

Tested with OpenFaaS Standard a long running sleep function which
went into a Terminating status. The function continued to
execute its sleep for the whole duration, whilst the new
replica came online and was ready in the meantime.

This is the same approach tested for the golang-http templates.

Signed-off-by: Alex Ellis (OpenFaaS Ltd) <alexellis2@gmail.com>