fix: ensure test layer is in the build DAG
Some builders will exclude layers that do not contribute to the final
target layer, meaning the `ship` layer for these templates. This means
that the `test` layer is often optimized away by modern builders.

We add the `test` layer as a dependency of the `ship` layer to ensure it
is always run.

Signed-off-by: Lucas Roesler <roesler.lucas@gmail.com>